requirements.txt
streamlit>=1.35
pandas>=2.0
plotly>=5.0
python-dotenv>=1.0
